+++ tests/src/net_box.rs
@@ -59,7 +59,7 @@ pub fn execute() {
         .expect("IPROTO execute sql request fail");
 
     assert_eq!(result.len(), 1);
-    assert_eq!(result.get(0).unwrap().as_struct::<(u64, String)>().unwrap(), (6002, "6002".to_string()));
+    assert_eq!(result.get(0).unwrap().decode::<(u64, String)>().unwrap(), (6002, "6002".to_string()));
 }
 
 pub fn ping_timeout() {
@@ -105,7 +105,7 @@ pub fn call() {
     let result = conn
         .call("test_stored_proc", &(1, 2), &Options::default())
         .unwrap();
-    assert_eq!(result.unwrap().into_struct::<(i32,)>().unwrap(), (3,));
+    assert_eq!(result.unwrap().decode::<(i32,)>().unwrap(), (3,));
 }
 
 pub fn call_async() {
@@ -117,7 +117,7 @@ pub fn call_async() {
     assert_eq!(p2.wait().unwrap(), (50,));
     assert_eq!(p1.state(), State::Kept);
     let tuple: Tuple = p1.wait().unwrap();
-    assert_eq!(tuple.into_struct::<(i32,)>().unwrap(), (489,));
+    assert_eq!(tuple.decode::<(i32,)>().unwrap(), (489,));
 }
 
 pub fn call_async_error() {
@@ -182,7 +182,7 @@ pub fn eval() {
     let result = conn
         .eval("return ...", &(1, 2), &Options::default())
         .unwrap();
-    assert_eq!(result.unwrap().into_struct::<(i32, i32)>().unwrap(), (1, 2));
+    assert_eq!(result.unwrap().decode::<(i32, i32)>().unwrap(), (1, 2));
 }
 
 pub fn eval_async() {
@@ -282,7 +282,7 @@ pub fn get() {
         .unwrap();
     assert!(output.is_some());
     assert_eq!(
-        output.unwrap().into_struct::<S2Record>().unwrap(),
+        output.unwrap().decode::<S2Record>().unwrap(),
         S2Record {
             id: 16,
             key: "key_16".to_string(),
@@ -300,7 +300,7 @@ pub fn select() {
     let result: Vec<S1Record> = space
         .select(IteratorType::LE, &(2,), &Options::default())
         .unwrap()
-        .map(|x| x.into_struct().unwrap())
+        .map(|x| x.decode().unwrap())
         .collect();
 
     assert_eq!(
@@ -332,13 +332,13 @@ pub fn insert() {
     let insert_result = remote_space.insert(&input, &Options::default()).unwrap();
     assert!(insert_result.is_some());
     assert_eq!(
-        insert_result.unwrap().into_struct::<S1Record>().unwrap(),
+        insert_result.unwrap().decode::<S1Record>().unwrap(),
         input
     );
 
     let output = local_space.get(&(input.id,)).unwrap();
     assert!(output.is_some());
-    assert_eq!(output.unwrap().into_struct::<S1Record>().unwrap(), input);
+    assert_eq!(output.unwrap().decode::<S1Record>().unwrap(), input);
 }
 
 pub fn replace() {
@@ -363,14 +363,14 @@ pub fn replace() {
         .unwrap();
     assert!(replace_result.is_some());
     assert_eq!(
-        replace_result.unwrap().into_struct::<S1Record>().unwrap(),
+        replace_result.unwrap().decode::<S1Record>().unwrap(),
         new_input
     );
 
     let output = local_space.get(&(new_input.id,)).unwrap();
     assert!(output.is_some());
     assert_eq!(
-        output.unwrap().into_struct::<S1Record>().unwrap(),
+        output.unwrap().decode::<S1Record>().unwrap(),
         new_input
     );
 }
@@ -403,7 +403,7 @@ pub fn update() {
     assert_eq!(
         update_result
             .unwrap()
-            .into_struct::<S1Record>()
+            .decode::<S1Record>()
             .unwrap()
             .text,
         "New"
@@ -411,7 +411,7 @@ pub fn update() {
 
     let output = local_space.get(&(input.id,)).unwrap();
     assert_eq!(
-        output.unwrap().into_struct::<S1Record>().unwrap().text,
+        output.unwrap().decode::<S1Record>().unwrap().text,
         "New"
     );
 }
@@ -461,13 +461,13 @@ pub fn upsert() {
 
     let output = local_space.get(&(1,)).unwrap();
     assert_eq!(
-        output.unwrap().into_struct::<S1Record>().unwrap().text,
+        output.unwrap().decode::<S1Record>().unwrap().text,
         "Test 1"
     );
 
     let output = local_space.get(&(2,)).unwrap();
     assert_eq!(
-        output.unwrap().into_struct::<S1Record>().unwrap().text,
+        output.unwrap().decode::<S1Record>().unwrap().text,
         "New"
     );
 }
@@ -490,7 +490,7 @@ pub fn delete() {
         .unwrap();
     assert!(delete_result.is_some());
     assert_eq!(
-        delete_result.unwrap().into_struct::<S1Record>().unwrap(),
+        delete_result.unwrap().decode::<S1Record>().unwrap(),
         input
     );